requirements.txt
requests==2.32.4
pytest==8.4.0
pytest-html==4.1.1
python-dotenv==1.1.0
jsonschema==4.24.0
pylint==3.3.7
